Going with doubles, working implementation of Gauss Jordan Algorithm

+++ include/gj_utils.hpp
@@ -7,19 +7,19 @@ namespace GJ_Utils{
         private:
         bool owns_data;
         public:
-            float * data;
+            double * data;
             int rows,cols;
             Matrix(int rows, int cols);
-            Matrix(float* externalData, int rows, int cols): rows(rows), cols(cols), data(externalData), owns_data(false) {}
+            Matrix(double* externalData, int rows, int cols): rows(rows), cols(cols), data(externalData), owns_data(false) {}
             ~Matrix();
-            float& at(int row, int col);
+            double& at(int row, int col);
             void print();
     };
 
     class S_Matrix : public Matrix {
         public:
             S_Matrix(int size) : Matrix(size, size) {}
-            S_Matrix(float* external_data, int size) : Matrix(external_data,size, size) {}
+            S_Matrix(double* external_data, int size) : Matrix(external_data,size, size) {}
             void fill_random_L();
             void fill_random_U();
             S_Matrix times(const S_Matrix *m2);
@@ -29,7 +29,7 @@ namespace GJ_Utils{
     class GJ_Matrix : public Matrix {
         public:
             GJ_Matrix(S_Matrix* matrix);
-            GJ_Matrix(float* allocated,S_Matrix* matrix);
+            GJ_Matrix(double* allocated,S_Matrix* matrix);
             void print();
             S_Matrix get_right_side();
             S_Matrix get_left_side();
